Fight (#52)
* WONDROUS_ITEM

* showItems

* Itemall

* itemReact_neDziala

* equip_exp

* Inventory

* itemsDTO

* butItems

* WondrousItems

* container

* sendRankToDB

* updatedInventory

* postNeDziala

* postItemOk

* soldi

* actualGold

* attackOne

* archi

* attackOneOrTwo

* list of attacks

* attacksDTo

* attacksNew

* fromDBToReactAtt

* attacksOk

* reload

* newArmorAddInJava

* enchantmentButton

* functionName

* equalProblem

* enchantmentAndMaterial

* armorDropDown

* someSpells

* letteraC

* letteraD

* addSpellToBook

* docker-compose

* funziaJava

* dockerfileNo

* backendDocker

* loadApi

* reactToAws

* styleRPG

* finoAClassi

* guiRPG

* centred

* finoAFeats

* correctFeats

* showChar

* deleteIdChar

* ShowCSS

* saving

+++ api/src/main/java/pl/kolendateam/dadcard/characterCard/repository/CharacterRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.kolendateam.dadcard.characterCard.entity.Character;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<Character, Short> {
-    
-}
+public interface CharacterRepository
+  extends JpaRepository<Character, Integer> {}
